requirements.txt
PyQt5
psycopg2-binary
geopy